fix: refactor modal into separated utils/hooks and add inner tabs
